+++ Cargo.toml
@@ -37,6 +37,8 @@ libra-config = { path = "tools/config" }
 libra-framework = { path = "framework" }
 libra-genesis-tools = { path = "tools/genesis" }
 libra-query = { path = "tools/query" }
+libra-rescue = { path = "tools/rescue" }
+libra-storage = { path = "tools/storage" }
 libra-smoke-tests = { path = "smoke-tests" }
 libra-tower = { path = "tools/tower" }
 libra-types = { path = "types" }
