+++ tools/cli/Cargo.toml
@@ -21,6 +21,8 @@ libra-config = { workspace = true }
 libra-framework = { workspace = true }
 libra-genesis-tools = { workspace = true }
 libra-query = { workspace = true }
+libra-rescue = { workspace = true }
+libra-storage = { workspace = true }
 libra-txs = { workspace = true }
 libra-types = { workspace = true }
 libra-wallet = { workspace = true }
