+++ tools/rescue/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rescue"
+name = "libra-rescue"
 version = "0.1.0"
 
 # Workspace inherited keys
